Don't skip client stream 1 on SPDY/3.1.
This breaks connectivity to google.com's SPDY/3.1 servers.
